Adding 'keydown' Event Listener
A 'keydown' event listener was added on the 'window' level
leveraging the event bubbling.

At this stage the audio object can be played and the sound works even
if the user pressed the same button multiple times without any delays.

What should be corrected is the fact that the class 'playing' stays on.