CSS Rules - Adding Rules for body and Attribution div elements
The body rules are adjusted as my HTML layout is different since I
added attribution for the background image.

The proper CSS rules were added for the div with the attribution ID,
in such a way that the user should scroll down a little for the
attribution to be revealed.