Solution 4: Array.prototype.reduce()
The reduce() method accepts a user-supplied callback function which
is called with the following arguments:
--> accumulator : the value resulting from the previous call to callbackFn.
  |--> On the first call its value is "initialValue", otherwise array[0].

--> currentValue : the value of the current element.
  |--> On the first call its value is array[0], should initialValue is specified,
       otherwise its value is array[1].

Thus, in this example, since we want the years lived in total by all inventors,
the initialValue should be 0.

Otherwise, a weird result will be printed, "[object Object]8478675970898137509080".
When the reduce() method is called without an initialValue, the first element of the array
used as argument to the sum argument, which is an object.
Thus, in the second interation the yearsLived are added to the string representation of the object, "[object Object]".
Therefore, all the years lived by the inventors are concantanted as strings.