Solution 7: Alphabetical Sorting by Last Name
When sort() method is called without arguments,
the array elements are converted to strings, then
sorted accoridng to each character's Unicode code
point value.

Since, the strings in the people array follows the
pattern, "lastName, firstName", it is not necessary
to pass a callback function.

For this challenge, after the names have been sorted,
some logic was applied to display the full name
without the command between last name and first name.