Solution 8: Summing Up Instances of Elements
The reduce() method can be used to count the instances of a
particular element in the array.

The result will be returned in the format of:
{ elementA: instancesA, elementB: instancesB, ..., elementN: instancesN}

For this to happen, the initialValue parameter of reduce() method should
accept "{}" as argument.

Furthermore, whether the object has a particular property must be checked.
If the property does not exist, then we have to create the property with
an initialization value of zero (0). Otherwise, an "undefined" value will
be increased by one which leads to "NaN".