Solution 4: Array.prototype.findIndex()
The findIndex() method returns the index of the first element
in an array that satisfies the provided testing callback
function.

In this challenge, since we want to delete an element from
the array, the target's element index is needed.

Finally, there are two ways to delete the element from the array:
--> Using the splice() method, which mutates the original array
  |-> comments.splice(commentIndex,1);

--> Combining the spread operator and slice() method to create a new array.
  |-> newArray = [ ...comments.slice(0, commentIndex),
                   ...comments.slice(commentIndex + 1)
                 ];