CSS Rules: Facelifting
The border property was removed which was acting as a visual to observe
the changes.

"flex:5" was added to the .panel.open instructing the element to grow
five time in relation to its neighbor flex items.

translateY(0) was added to display the first and last <p> elements
when the <div> element acquire the "open-active" class.