JS Code: Adding Event Listeners
The HTML file was modified to include the JS file.
In the JS file, code was written to acquire all the elements that
have the class ".panel". At first, the ".open" class is toggled
when the user clicks on the image. Then, when the transition is
finished another class is toggled, ".open-active".